Replace explicit version number with versioneer
Previously (prior to #228), we used the [PBR][PBR] package with an
explicit version number in setup.cfg. This enables the 'pre-versioning'
mode of managing version strings (see [features][PBR-versioning]) which
raises an error if the inferred version from Git tags is greater than
explicit version in setup.cfg. This prevents developers from doing an
editable install via `pip install -e .` (see #129). This issue was fixed
implicitly in #228 when PBR was removed from the build dependencies.

This change is a feature in that it implements dynamic version strings
from Git tags via [versioneer][versioneer] so that an explicit version
number in setup.cfg is not required. We choose [versioneer][versioneer]
so that version strings were consistent across `pip` and `conda`
installations. This was not possible with other tools such as [PBR][PBR]
and [setuptools_scm][setuptools_scm].

Fixes #129

[versioneer]: https://github.com/python-versioneer/python-versioneer
[PBR]: https://docs.openstack.org/pbr/latest/index.html
[PBR-versioning]: https://docs.openstack.org/pbr/latest/user/features.html#version
[setuptools_scm]: https://setuptools-scm.readthedocs.io/en/latest/