Disable module load, unload, ... invocations via wrapper
This change disables module commands that modify the environment via a
wrapper when running a custom build script. This avoids preprocessing
the custom build script for module statements which may produce an
invalid bash script. This also allows for module commands that do not
alter the environment to be executed as we only disable a subset of
subcommands in the wrapper.

Fixes #249