Unset LDFLAGS and CMAKE_PREFIX_PATH variables
CABLE is linked against incorrect libraries for netcdf and MPI when
running benchcab (v4.0.2) from the hh5 conda environment. The issue is
due to environment variables being set which affect the behaviour of the
build, notably LDFLAGS and CMAKE_PREFIX_PATH, which point CMake to find
the netcdf and MPI libraries installed in the current conda environment.
This change unsets these variables so that CMake finds the appropriate
libraries which get loaded in as modules.

Fixes #279