Log status of fluxsite and comparison runs
This change introduces a State object as a minimal way of having state
persist between separate processes. This is necessary for correctly
showing the status of fluxsite and comparison runs as these tasks are
run inside child processes which do not share the same data structures
in the parent process.

Fixes #180