Remove module load/purge statements from PBS job (#330)
Loading modules prior to launching benchcab may break the environment
(as evident in #329 for the xp65 container environments).

Modules do not actually need to be loaded when running CABLE as CMake
sets the RPATH on the executable to point to all the library
dependencies, so it is safe to remove these module statements from the
job script.

Fixes #329